Removed some more things
